requirements.txt
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gym==0.25.2
gym-notices==0.0.8
gymnasium==0.29.1
highway-env==1.8.2
idna==3.4
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
-e git+https://github.com/HumanCompatibleAI/overcooked_ai.git@83f51921869f25ddf5153aa2742b4e1d4da5e6e9#egg=overcooked_ai
packaging==23.2
pandas==2.2.0
parso==0.8.3
pettingzoo==1.24.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.13
wcwidth==0.2.13
widgetsnbextension==4.0.9
